wip: Setup network layer (ktor)
