+++ data/src/main/java/com/android254/data/network/Constants.kt
@@ -1,0 +1,6 @@
+package com.android254.data.network
+
+object Constants {
+    const val BASE_URL = ""
+    const val EVENT_SLUG = ""
+}
