+++ data/src/main/java/com/android254/data/network/models/Responses.kt
@@ -1,0 +1,3 @@
+package com.android254.data.network.models
+
+data class Status(val message: String)
